refactor(scenario): move scenario type to existing file
