--- src/types/scenario.ts
@@ -1,9 +1,0 @@
-import {DocumentReference} from 'firebase/firestore';
-
-export interface Scenario {
-  id: string;
-  name: string;
-  baseLayerTypes: Array<string>;
-  balancedScorecardsRef: DocumentReference;
-  notesRef: DocumentReference;
-}
