+++ src/types/scenarios.ts
@@ -1,6 +1,14 @@
-export interface ScenarioMetaData {
+import {DocumentReference} from 'firebase/firestore';
+
+export interface Scenario {
   id: string;
+  name: string;
+  baseLayerTypes: Array<string>;
+  balancedScorecardsRef: DocumentReference;
+  notesRef: DocumentReference;
+}
+
+export interface ScenarioMetaData extends Pick<Scenario, 'id' | 'name'> {
   balancedScorecardsId: string;
   notesId: string;
-  name: string;
 }
